Fixing vectors to use 64-bit indexing

+++ src/discretization/DOF_Manager.cc
@@ -9,11 +9,11 @@ namespace Discretization {
 /****************************************************************
 * Get the entry indices of DOFs given a mesh element            *
 ****************************************************************/
-void DOFManager::getDOFs( const AMP::Mesh::MeshElement &obj, std::vector <unsigned int> &dofs, std::vector<unsigned int> ) const
+void DOFManager::getDOFs( const AMP::Mesh::MeshElement &obj, std::vector <size_t> &dofs, std::vector<size_t> ) const
 {
     AMP_ERROR("getDOFs is not implimented for the base class");
 }
-void DOFManager::getDOFs( const AMP::Mesh::MeshElementID &id, std::vector <unsigned int> &dofs ) const
+void DOFManager::getDOFs( const AMP::Mesh::MeshElementID &id, std::vector <size_t> &dofs ) const
 {
     AMP_ERROR("getDOFs is not implimented for the base class");
 }
